requirements.txt
langchain
langchain-google-genai
gradio
python-dotenv
pedantic